// src/interfaces/cadastroClienteInterface.ts
export interface CadastroInterface {
    id: number;
    nome: string;
    cpf: string;
    email: string;
    celular: string;
    dataNascismento: string;
    cidade: string;
    estado: string;
    pais: string;
    rua: string;
    numero: string;
    bairro: string;
    cep: string;
    complemento: string;
    senha: string;
}